fix(levm): call with no code in address   (#792)
**Motivation**

Noticed that when a call is executed on an address with no code and the
VM tries to read the code data, the default value is returned, 0, which
corresponds to the `STOP` instruction and halts the execution.

Right now we are pushing 0 to the stack and continuing with the
execution, this should be fixed.

**Changes**
- Update `generic_call` to return an `OpcodeSuccess`, this lets us to
return a `REVERT` or a `HALT` on all the call functions. Previously we
were always returning a continue.
- Added a `Revert` to the `ResultReason`.