fix(l2): gracefully stop `ProofDataProvider` (#851)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
When `Ctrl+C` is sent to the process, the `ProofDataProvider` listener
is still opened waiting for new connections, causing the process to not
finish.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Now the listener is launched in a new thread, and the previous thread
have a signal handler that sends a message under a `mpsc` channel to the
listener so it stops.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #809