feat(l1): add Assertoor support to localnet (#857)
**Motivation**
Be able to run a suite of tests over a localnet as part of the CI or
post merge

**Description**
Added assertoor to the localnet. To learn about assertoor, see:
https://ethpandaops.io/posts/assertoor-introduction/

As part of this task, I had to add the `web3_clientVersion` RPC
endpoint, which for now returns `ethereum_rust@0.1.0`. We can improve
this later.