feat(l1): add libmdbx features  (#875)
**Motivation**

`mdbx` cannot be compiled for the `rv32im` target (used in L2 prover) so
this introduces a feature to disable it in those cases, as we need to
import some `ethereum_rust` crates in the program.

This PR also adds (De)Serialization for types that we need to send to
the zkVM program (which are (de)serialized in a binary format)

**Description**

- Bump up `revm` version
- Set `features` to compile limbmdbx conditionally 
- Avoid `revm's` `default-features` &rarr;
https://docs.rs/crate/revm/14.0.3/features
- The `blst` feature from `revm` imported by `ethereum_rust-evm` was
causing some issues, so it was disabled with `default-features = false`.
- `mdbx` is compiled conditionally based on the `libmdbx` feature.
- added CI job for building the prover zkVM program.
- adds (de)serialization for types that need to be sent to the zkVM.

---------

Co-authored-by: Estéfano Bargas <estefano.bargas@fing.edu.uy>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>