feat(l2): use `ethereum_rust` dev as local L1 (#890)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

We need a special setup of this client for users to play around with a
local L1 devnet because regularly the execution client needs a consensus
counterpart to advance blocks. For example, when initializing a local
L2, we use a local L1 devnet that continuously produces blocks. Today
we're using `reth` but want to replace it asap.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

- Adds a `dev` (`ethereum_rust-dev`) crate with
- an `engine_client` implementation. An `EngineClient` is an HTTP client
that supports performing calls to the EngineAPI. It is used for
"advancing the blockchain".
- a `block_producer` module with the logic that makes the blockchain
"advance".
- `ethereum_rust-dev` is added as an optional dependency in
`cmd/ethereum_rust` and is only used under the `dev` feature flag, which
starts the block production if set.
- Adds a Dockerfile for running a `dev` client. This is useful for
tracking the client logs, as most L2s are run by a single command, and
if the client is run locally, its logs would interfere with the L2 ones.

**Side notes**

- I needed to change `eth_gasPrice` behavior a little bit. Before, we
were returning an error when someone requested the `eth_gasPrice` and
the only block in the blockchain was the genesis. Now that we can start
the client in a dev environment starting with only the genesis block,
we've updated the response to reply with the minimum gas price.