feat(levm): add transact validations (#921)
**Motivation**

This PR goal is to add validations in transact that are not done in
upper layers

**Description**

The validations are based in [yellow
paper's](https://ethereum.github.io/yellowpaper/paper.pdf) transaction
execution section (the 6th one). At the moment there is a lack of tests
of the Cancun fork.
Also, removes some constants about transactions intrinsic gas costs,
because we don't do those calculations in this layer.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #900