feat(l1): assertoor CI integration (#943)
**Motivation**

Assertoor integration as part of the CI

**Description**

This PR adds the [Assertoor
GHA](https://github.com/ethpandaops/kurtosis-assertoor-github-action) to
have a local devnet running with some checks as part of the CI. Most of
the time for the stability checks was related to consensus, after
removing those checks in a custom playbook created from
[`stability-check.yaml`](https://github.com/ethpandaops/assertoor-test/blob/master/assertoor-tests/stability-check.yaml)
we are able to run it as part of the CI. For assertoor to run with our
[fork of
ethereum-package](https://github.com/lambdaclass/ethereum-package) a
small
[commit](https://github.com/lambdaclass/ethereum-package/commit/811581ee8ab42639e4e9ef7c1197041689bfdfd1)
was needed to the `ethereum-rust-integration` branch.

I've also added some entries to the `paths-ignore` on the push rules.

_Note about time: Most of the time spent in this new job is not the
assertoor run, which takes ~3-5 mins, but the actual building of the
docker image, which take ~9mins before starting the tests. This is done
in parallel to the docker build in the CI, but now it's done 2 times.
Fortunately, this is not the longest run, since `Hive / Cancun Engine
tests (pull_request)` takes ~19m_

Resolves #952