feat(l1): in-memory cache database and new `EvmState` (#888)
**Motivation**

implement a memory cache database and expand the `EvmState` struct into
an enum. This is to take advantage of the block execution and validation
implementations of the L1 to use it for proving in L2 mode.

**Description**
- adds new `EvmState` enum containing two variants (store and cache)
- adds `ExecutionDB` database for storing only the relevant data of the
execution of a block (this is going to be passed into a zkVM for
proving)
- adds and modifies some `EvmState` functions and adapts the codebase to
those changes

---------

Co-authored-by: fborello-lambda <federicoborello@lambdaclass.com>
Co-authored-by: Federico Borello <156438142+fborello-lambda@users.noreply.github.com>