refactor(l2): add `-w` flag to cmds that send txs (#1036)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

Nowadays, all the commands that send transactions do not wait for
transaction receipts. If you run the same command multiple times the
same transaction with the same nonce is going to be sent to the node;
another problem is that the users have to perform additional steps to
make sure that their transaction was finalized or not.

As this is not an implementation problem but a misuse of the CLI, it'd
be good for users to also have the option to wait for their transactions
to be finalized.

**Description**

Adds a `-w` flag to the cmds `deploy`, `transfer`, `send`, `deposit`,
`withdraw`, and `claim-withdraw` which when set, waits for the
transaction sent to be finalized (a.k.a. wait for its receipt).