feat(core): in memory mempool (#1002)
**Motivation**

This PR aims to move the mempool (i.e. transaction pool) from being a
`libmdbx` table to an in-memory data structure.
This makes sense as the mempool is not something meant to be persisted
for a long time, but rather a temporary place for transactions for
(ideally) a few seconds/minutes until they're included in blocks or
discarded. If the node goes down and then back up for whatever reason,
mempool transactions will come back from the p2p gossip network anyway.

More concretely, though, the mempool being on disk was being a huge
bottleneck on the load tests we were running, which were sending
approximately 5k transactions per second to the node. When doing this
with the current on-disk mempool, the node would hang and completely
stop executing new blocks until the load test finished sending all
transactions, at which point it would resume execution. With this (very
naive and rushed) in-memory implementation, the node does not hang and
can keep executing blocks at a normal pace while also receiving and
storing all incoming transactions.

Things to fix in the future:

- Limiting the size of the mempool. We'd like to avoid the possibility
of an OOM due to too many transactions (note that this was also a
problem with the on-disk structure, running out of disk space instead of
memory though).
- Optimizing the way transactions are stored, filtered and picked to be
included on each block. To build a block, the node currently does the
following:

- Keep a data structure of all the mempool transactions, grouped by
sender address and sorted by nonce: for each sender address, we want
transactions ordered by nonce in increasing order to know what the next
transaction is for an account quickly. This is the `txs` field of the
`TransactionQueue` struct right now.
- Keep a data structure of all the `head` transactions sorted by fee tip
in decreasing order, i.e., a list that has, for each sender address, its
first transaction in terms of nonce, with the highest paying one first
and so on (this is just a subset of the above where we just keep the
first transaction for each key on a list and then sort that by fee).
This is the `heads` field of the `TransactionQueue`.
- While there's still gas left in the block, take the first transaction
from the head transactions list and execute it; for that transaction's
sender, take their next transaction (in terms of nonce) and add it to
the head transactions list, making sure it remains sorted after
insertion. Repeat until the block is full.

This ensures the node builds a block with all valid transactions (i.e.
no invalid nonces) and with the highest possible fee for them. However
right now there's a lot of copying going on because the mempool is kept
separately from the `TransactionQueue`; to build the `TransactionQueue`
the builder first takes each mempool tx (stored on the `mempool` field
of the `Store`) and passes it through a filter, then copies the remains
of that filter to the Queue and builds the sorted head transactions list
from there.

My impression is that a lot of this copying and duplicating of
transactions can be removed.