fix(levm): fix calldataload (and add sload gas consumption) (#1068)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
Calldataload was giving problems the way it was implemented. Changed
implementation and now it doesn't produce an error in execution.

Description:
- Fixes CALLDATALOAD, before it only worked in the ideal scenario where
the data read is 32 bytes long, but that does not always happen.
Sometimes calldata is less than 32 bytes and other times the offset is
so high that from there to the end there are not 32 bytes. -> Solution
was to fill the rest of the bytes with zeros.
- Added SLOAD opcode gas implementation (slightly modified from #1053)
because the function I was testing in the solidity contract requires
that to be implemented.


Results:
- Now sending a transaction that calls the function `number()(uint256)`
in a basic solidity contract seems to work properly.
- As a side effect of fixing that it also works to call the function
`increment()` of the basic contract.

What's left:
- Gas refunds haven't been implemented in SSTORE yet so if I were to set
the storage to 0 then it wouldn't work properly