feat(levm): prepare levm-ethereum-rust-integration for merge (#1064)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

Eventually, we want to stop using revm and replace it with levm. The
integration is a work in progress in this PR
https://github.com/lambdaclass/lambda_ethereum_rust/pull/1029.

This PR aims to be the first step. It aims to prepare the current
codebase so that both implementations can coexist while levm continues
to be developed and does not disturb the rest of the development. When
the time to remove revm comes, it'd be easier with this proposal since
it'd require removing the code under the #[cfg(not(feature = "levm"))]
and after that, to remove the feature flag definitively.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

- Adds the `levm` feature flag to the `vm` and `blockchain` packages,
and to the `ethereum_rust` binary.
- In `vm`, add `execute_block` implementation that uses `levm` and use
it when `levm` feature flag is set. Use the original implementation if
not.
- In `blockchain`, add `add_block` implementation that uses `levm` and
use it when `levm` feature flag is set. Use the original implementation
if not.

---------

Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>
Co-authored-by: JereSalo <jeresalo17@gmail.com>