feat(levm): ethereum rust integration (#1029)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

Eventually, we will stop using revm and replace it with levm.

This PR aims to be the first step. It aims to prepare the current
codebase so that both implementations can coexist while levm continues
to be developed and does not disturb the rest of the development. When
the time to remove revm comes, it'd be easier with this proposal since
it'd require removing the code under the #[cfg(not(feature = "levm"))]
and after that, removing the feature flag definitively.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

- Adds the `levm` feature flag to the `vm` and `blockchain` packages,
and to the `ethereum_rust` binary.
- In `vm`, add `execute_block` implementation that uses `levm` and use
it when `levm` feature flag is set. Use the original implementation if
not.
- In `blockchain`, add `add_block` implementation that uses `levm` and
use it when `levm` feature flag is set. Use the original implementation
if not.

---------

Co-authored-by: Juani Medone <juan.medone@lambdaclass.com>
Co-authored-by: maximopalopoli <mpalopoli@fi.uba.ar>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>
Co-authored-by: ilitteri <ilitteri@fi.uba.ar>