fix(l1): fixing transaction spammer delay and nonce error on estimation (#1092)
**Motivation**

This PR solves all the issues discovered while starting the tx spammer
pointing to our node in the kurtosis setup. Right now it even work with
the default parameters (spamming 100~200 tx per block)

**Description**

There were a bunch of issues generating the tx to be both delayed and
ultimately stop after the first one, they were the following:
- We didn't had into account the base fee when there were no txs in the
blocks we proposed, as happens starting from genesis and being the only
one targeted by the spammer. Now we default first to the base fee and
then to the min gas limit. This was causing txs to wait until the 90th
block to start being added.
- We were not taking into account pending transactions for nonce
calculation when the pending tag was sent
- This was making subsequent transactions for the same account to halt
the spammer if some of them were still in the mempool.
- During the simulate tx of estimate gas we were defaulting the nonce to
0 if not provided. This was crashing the spammer on subsequent tx calls
for the same account to estimate gas, now we have an Option for the
nonce as part of the GenericTransaction.
 
Resolves #1080
Resolves #1081
Resolves #1083
Resolves #1116