fix(levm): fix revert cases for contracts to run correctly (#1090)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
Fix errors encountered when trying to execute solidity contracts with
LEVM.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- Fixes gas consumption in transaction report -> This fix enables us to
correclty execute a contract that triggers an `OutOfGas`
- Balance substraction is now done after executing the contract code. It
will be only performed if the execution result was successful. -> This
fix enables us to run a contract that involves a transfer of value and
reverts.
- Fixes problems when executing a contract that calls another one

Disclaimer:
- The PR could be improved because the code is pretty messy but I think
there are other priorities right now so I didn't focus on coding it the
best way possible. There are a few `unwraps` out there

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Juani Medone <juan.medone@lambdaclass.com>
Co-authored-by: maximopalopoli <mpalopoli@fi.uba.ar>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>
Co-authored-by: ilitteri <ilitteri@fi.uba.ar>