feat(l2): prover's initial state validation (#1013)
**Motivation**

We need to verify the of initial final state data during the block
execution performed by the L2 zkVM. Currently we can't implement a
function to decode and verify a Trie proof, this problem is being solved
by #1076.

**Description**

This PR started with the idea of verifying both initial and final state
proofs, but now the idea changed to only verify initial proofs and
compute a pruned tree in the zkVM for validating final state data.
Because of this most of the implementation was deleted from this PR and
only verification of initial state remained. Pruned Trie computation
will be implemented in a future PR.

- added trie proof verification mockup
- implemented `StateProofs` and Trie proof retrieval when the
`ExecutionDB` is being created.
- add proof verification in the prover program

---------

Co-authored-by: fborello-lambda <federicoborello@lambdaclass.com>
Co-authored-by: Federico Borello <156438142+fborello-lambda@users.noreply.github.com>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>