refactor(l2): simplify transactions creation (#1142)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
We need a more centralized way to create and send transactions, so it's
more difficult to make a mistake creating one manually

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
This PR makes a big refactor on how transactions are built. EthClient's
`build_X_transaction` and `send_X_transaction` are used to build and
sign/send transactions respectively.
`libsecp256k1` was replaced with `secp256k1` to unify libraries across
crates.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #1024