feat(levm): initialize cache with sender, recipient and coinbase accounts (#1114)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

As stated in [evm.codes](https://www.evm.codes/about#access_list), this
PR initializes the cache with the corresponding addresses. (Without
implementing yet the concept of an Access List transaction)
<!-- A clear and concise general description of the changes this PR
introduces -->
Notes: Our `Cache` struct replaces the need for a `Substate` that keeps
warm addresses and storage slots because we know that if an account or
storage slot is in our cache then it is already warm.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #1088

---------

Co-authored-by: Juani Medone <juan.medone@lambdaclass.com>
Co-authored-by: maximopalopoli <mpalopoli@fi.uba.ar>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>
Co-authored-by: ilitteri <ilitteri@fi.uba.ar>