fix(l2): fix zkVM performance test (#1202)
**Motivation**

This test uses the `test_data/genesis-l2.json` and
`test_data/l2-loadtest.rlp` files. The genesis file was changed in
#1158, breaking the test because now the genesis block hash changed and
so the second block's parent isn't present in the chain anymore. The
test only executed in the CI whenever a file in `crates/l2/prover/` was
touched, leaving the breaking change unnoticed.


Needed for #1133

**Description**

- bring back the old genesis file as `test_data/genesis-l2-old.json` (a
better solution would be to regenerate`test_data/l2-loadtest.rlp`).
- change the test's CI job to always execute.