feat(levm): ensure returned errors are correct (#1225)
**Motivation**

Currently in EF tests we don't verify that the error returned by levm
matches with the expected, the goal is to verify it, and print an error
if it does not match.

**Description**

This PR includes:
- The logic of verifying the exception is the expected, in
`exception_is_expected()`
- The print of the error (or errors) that the EVM may throw in case the
error does not match or in case it does not return an error and one is
expected.