fix(l1): fix engine api v3 hive test related to latest valid hash and errors on ForkchoiceState (#1233)
**Motivation**

Fix the `engine-cancun` suite tests that we are able to fix with our
current implementation (without v1, v2 or reorg/syncing working as
expected)

**Description**

After a couples of back and forths the actual solution for both errors
are pretty simple, in the case of the latest valid hash it was needed to
send the parent when the evm fails after the parent was checked. In the
case of the ForkchoiceState we were sending an incorrect code we needed
to use the RpcError for FokchoiceState that we already had implemented.

This pass **31** new tests. On regards the Engine Cancun step in the CI,
previously it ran 38 tests, now it runs 149.

**Summary of the outstanding failing test in engine-cancun**

- 1 test (Blob Transaction Ordering) expect 6 blobs but got 5
- 2 test (Invalid NewPayload, ReceiptsRoot) Expected INVALID but got
VALID
- 4 test (Sidechain Reorg|TransactionRe-Org|Re-Org back into canonical
Chain) fails on Reorg or latest valid hash
- 15 tests ({exec_method}V[3|2]) are related to Shanghai (V2)
compatibility (v2 method not found)
- 13 tests(Fork ID: Genesis|Pre-Merge) are also related to previous
versions, (v2|1 method not found)
- 2 test are related to pooled blobs and depend on DevP2PRequestPooled
- 24 tests (Invalid Missing Ancestors) are related to syncing and
timeout

Resolves #1235