refactor(l2): restart processes with the proposed API (#1234)
**Motivation**

The error handling should be the same for all processes, unwraps have to
be handled with the `?` operator so that the api can restart the process
if any `Err` is raised.

**Description**

- Use the same API proposed by the PR #1230, removing the unwraps. 
  - #1168
- When we have an error we "restart" the process, this handles the `not
committed block` error.
  - #1165
- The order of the require statements in the `verify()` function of the
`OnChainProposer` contract was changed to throw the correct error. The
error handling with the proposed API does the rest.
  - #1164.

Closes #1168
Closes #1165 
Closes #1164

---------

Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>