ci(core): add back docker build workflow (#1295)
Reverts lambdaclass/ethrex#1286

It turns out adding `workflow_run` for Hive and Assertor was working
after all (it took some time, wtf github actions). Do I re-added it.

Merge queue only runs when the PR is not synced with main, so it does
not work as a guarantee. Instead I moved most checks to be run as part
of `push`, to at least be able to pin point if something breaks.