ci(core): standarize all CIs with required checks (#1334)
**Motivation**
Required checks in Github works in a weird way, it is based on the job
name. So we will define a couple of required checks:
- Lint
- Test
- Build Docker
- Integration Test

Each team will have one or more workflow, prepended with "ci_". There
they need to define all the above jobs. If there are multiple jobs with
the same name, they are all required.

For the merge queue, we keep it minimal and aim only to avoid breaking
changes that should be caught by the Rust compiler or linter. We don't
want to run heavy integration tests since those will be run on the push
event anyway.

**Description**
- Standardized ci workflows
- Standardized job names
- Simplified Hive to don't overuse makefiles (we don't need the build
capabilities)
- Added Integration Test job
- Re-added ci-skipped workflow for documentation changes.