fix(levm): modify transfer timing (#1359)
**Motivation**

Previously the value was added to receiver a subbed to sender twice with
a delegatecall operation, this PR fixes it. Also, the transfer of the
value was done when the transaction was confirmed, disallowing self
recursive contracts.

**Description**

This is done by using a should_transfer_value flag, that performs the
value transfer when it is activated, and transferring before the execute
and reverting if result is revert.

---------

Co-authored-by: ilitteri <ilitteri@fi.uba.ar>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>