feat(levm): improve LEVM EFTests (skip tests, disable spinner flag, execute individual tests) (#1394)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- Running tests is pretty slow but the bottleneck stands in a few tests
only, this functionality will allow the CI to run in less time and for a
person to execute most of the tests quickly.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- It skips `loopMul.json` arbitrarily because it is the test that causes
the biggest bottleneck. It is hardcoded just because if the person
that's testing forgets to skip that test, they are going to regret it.
- It allows the user to make use of the flag `--skip` to skip tests and
directories.
- It allows the user to send multiple tests to both flags --test and
--skip like this: ... --skip folder1,test1,folder2...

**Additional Features**
- It adds flag flag disable-spinner, if set, the main spinners will stop
and it will replace the "spinner prints" for `println!`, useful for CI.
(not yet)
- Adds functionality to send individual tests to `--test ` instead of
only folders. Useful for fixing tests

**CI**
- Improves readability and speed of ci, now it is run with `make
run-evm-ef-tests-ci`

**Example Usage**
`cargo test -p ef_tests-levm --test ef_tests_levm -- --tests
intrinsic.json,add11.json,stChainId --skip chainId.json
--disable-spinner`
This command runst tests `intrinsic.json`, `add11.json` and the ones in
the folder `stChainId`. But this also skips the test `chainId.json`,
this one is inside `stChainId` and therefore only the rest of the tests
in that folder will execute.

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>