refactor(levm): memory module (#1403)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

- The `Memory` implementation was similar to type aliasing a `Vec<u8>`.
The current implementation unnecessarily adds an extra layer of
indirection.
- The current `Memory` implementation has some bugs and not all its
methods are panic safe.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

- `Memory` is now a type alias for `Vec<u8>`.
- All the `Memory` "methods" are implemented as functions in the
`memory` module and called taking advantage of the namespacing (e.g.
`memory::load_word(...)`).
- All the `Memory` functions are panic safe.
- Memory access and expansion cost functions were refactored and moved
to the `memory` module. As a side note, in a future refactor we will be
updating the gas cost type from `U256` to `usize`, this PR kind of
introduces this notion, but it turns the `usize`s into `U256` wherever
is necessary for backward compatibility the implementation.

---------

Co-authored-by: Maximo Palopoli <96491141+maximopalopoli@users.noreply.github.com>
Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>
Co-authored-by: maximopalopoli <mpalopoli@fi.uba.ar>