fix(levm): codecopy: return a 0 array if the offset is larger than the bytecode length.  (#1465)
**Motivation**
Currently, the offset is cast to usize immediately . This causes an a
`VeryLargeNumber` number error, and thus the memory is never updated.

<!-- Why does this pull request exist? What are its goals? -->

**Description**
Now, we check if the offset is larger than the bytecode itself. If that
is the case, then we simply return an array of `size` many zeros.

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

This won't compile because it depends on the changes made here:
https://github.com/lambdaclass/ethrex/pull/1458