fix(levm): don't update memory in extcodecopy if offset is larger tha… (#1471)
**Motivation**
Postpone `extcodcopy`'s cast to usize until it's checked that it can
fit.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
extcodecopy receives as input an offset. That offset can be of size
u256. This need to be casted to usize to index the code it want to
access. However, it is done too early; which causes a problem when the
offset is larger than the code itself.
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes: #1469