fix(levm): remove consumed gas from environment (#1487)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- We only want to keep track of the consumed gas in the current
callframe. We don't need to keep track of it in the environment.
- This change will allow us to implement accurate behavior related to
gas consumption in XCALL and CREATE opcodes.

Note that the gas consumed by the first callframe will indeed be the gas
consumed by total execution, because the gas consume by child CallFrames
will be added to the "father". (It is not as straightforward but that's
the idea).
This change doesn't fix nor break any test, but will be useful for
following changes

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number