docs(levm): documentation (#1486)
**Motivation**

We figured out that a common problem arises when offsets and indexes in
U256 are cast to usize early on in the opcodes execution. These can
sometimes cause errores if the value is larger than usize's
representation range.

The best way we figured out to deal with this problem was to delay the
usize cast as much as possible (or at least until we check that the
offset represents a valid position).

However, we did not have this methodology written down in the
repository. This PR adds documentation describing this approach, which
should be especially useful for newcomers to the the project.

**Description**
Add a `faq.md` file where this approach is described.

Closes #1482

---------

Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>