fix(levm): move code deposit cost of CREATE (#1513)
**Motivation**

We were miscalculating the `gas` of the `create()` because we always
added the `code_deposit_cost` to the `gas`. We should do that only when
the contract is successfully created.

**Description**

- Change name from `create()` to `generic_create()`
- Remove the `code_deposit_cost` from`compute_gas_create()`, that is
responsible for calculating the base gas of the create. This is now done
in case that `TxResult::Success` in `create()`.
- Check the prefix of the deployed code. If it's `0xEF`, we return an
error.
- We were assigning the bytecode of the new account as soon as it was
created. I removed that step and now create it with empty bytecode. Now
it is assigned only when it completes successfully.

---------

Co-authored-by: Leandro Serra <leandro.serra@lambdaclass.com>