fix(levm): change blobhash fee calculation  (#1509)
**Motivation**
The way we calculate the blob fee is not following the specification;
which meant over charging the sender.
Also, this changes the way we return a 0 in case the index is not valid
(that is, when the index is larger than the blob hash array itself).



**Description**

We are using the `tx.max_fee_per_blob_gas` when the specification uses
the `base_fee_per_blob_gas`. For this, we use the already available
function `get_base_fee_per_blob_gas`.


Also, now a 0 is pushed to the stack when the index is larger than usize
when it reality it should check against the amount of blobs. The
previous implementation achieved this, but this new implementation is
more explicit about _when_ it does it.

Closes https://github.com/lambdaclass/ethrex/issues/1508

---------

Co-authored-by: JereSalo <jeresalo17@gmail.com>
Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>